small, fast, straightforward Class implementation
This implementation gets rid of all of the run time overhead which comes
with e.g. mootools Class. The API is _almost_ the same, but Class itself
is not a constructor anymore, but merely a functions which returns the
constructor. This has the advantage of making the 'instanceof' operator
work as before. It takes away the need of having a new instanceOf
function like mootools does. API example in class.html.