Restructured several widgets to use scheduled redraw()
Instead of mutating the DOM immediately when an option changes, the
set() method merely marks the option as invalid and schedules the redraw
method with the dom scheduler. Once the redraw method is called, all
changes are applied to the DOM. Many Widgets have not been refactored
completely, this is still an experimental.
This improves initializiation times of the e.g. levelmeter drastically.
For instance, showcase#LevelMeter loads about 30% faster.
However, there is still a long way to go. Currently individual parts of
the redraw() trigger relayouts unnecessarily. Therefore, there is still
much to be gained when switching to using Generators. However,
refactoring the redraw() routines will be the first important step.