Tooltip: remove keep_inside
keep_inside kills performance, and in case of the tooltip it essentially
triggers a style recalculation on _every_ mousemove. We either have to
come up with a better positioning logic, which avoid calculating the
effective tooltip size (e.g. by moving it with translate()) or live with
a tooltip which moves out of the window.