Container: fixed hide/show-transitions
The hide/show states were set by a callback with timeout, to make
sure that the transition was finished. This could sometimes clash with
a call to hide/show. When the corresponding callback would fire _before_
the redraw happened, the state could end up being set to the wrong
state.

Fixes #40