TK.Scale: new automatic scale rendering
This change introduces a new algorithm for automatically rendering
a scale into the available pixel space. Apart from drawing scales
differently (see below), the main new feature is that it can correctly
avoid drawing those dots, which overlap with labels. This makes
automatic scales much more useful.

How the new algorithm works:

Dots and labels are (in principle) positioned in the same way. Given a
mainimum gap in px between two elements on the scale, and an array of
levels (i.e. possible step sizes in the value range), the algorithm
tries to place elments while favouring the highest possible values.
More specifically, it will start using the highest level on a given
interval and try to fit all those elements into the range, which are not
closer than the minimum pixel gap to any of its neighbors. After placing
all elements on the given range this process will be repeated
recursively on each resulting subinterval for the next smaller level.
This will be continued until no more elements fit into any interval.
If 'avoid_collisions' is true, it will remove all those dots, which
collide with labels. If 'show_markers' is true, it will place special
dots (with class 'toolkit-marker') at each label position.

The above algorithm is applied to the interval from 'base' to 'min' and
'max'. This only makes a difference if the levels are not divisors of
the full interval and each other.