Widgets: moved resize handling into the scheduler
Widgets now keep track of whether or not they were resized. This is
triggered by a new method trigger_resize(), which bubbles down the full
widget tree. The actual resize() method is then run by the scheduler in
priority 0. This replaces the logic, which was previously taken care of
by surrounding containers. This refactoring also makes sure that
children added into the widget tree are resized once. This initial
resize is delayed until _after_ the next rendering frame, to make sure
that widgets are actually in the DOM when resize runs.