New resize logic again
The actual resize now always happens _after_ all other rendering is
done. While this might be unnecessary in many situations, in which
the actual drawing is not necessary to run before the resize, it
simplifies many other things. Most importantly, it naturally solves the
problem of resizing for the initial render. In those cases it is
important to apply _all_ changes before rendering, to make sure that the
actual resize measures everything correctly.