ValueKnob: renamed the 'label' option
For some future release all options of named subwidgets are supposed to
become accessible through the <widget>.<option> syntax. This is not
implemented anywhere yet, but we do not want to introduce another option
using the old nameing.
Also, 'label' clashes with another option of the knob, which is why it
needs to be renamed, anyway.