DragValue: reenabled preventDefault in mousedown
The call to preventDefault in mousedown has important reasons, such as
disabling text selection. Unfortunately, one side effect is that in
_some_ browsers (chrome) mouse events from outside of an iframe are not
delivered anymore. The workaround here is to assume that if a user
wanted to stop dragging, the pressed buttons will change. Checking for
that in mousemove will at least prevent most 'bad' behavior.

Chrome has now fully implemented the Pointer Events API in version 55.
This API allows better control of pointer capture. Firefox has support
coming. The right plan now is this: build a minimal abstraction for
pointer capture and use that everywhere.