DragValue: fixed blind_angle calculation
Since the switch to DragCapture the blind_angle calculation was broken.
Previously, the drag distance was added to the value continuously. Since
DragCapture was introduced the value change always depends on the
distance since the start of the dragging. This meant that the
calculation had to be fixed. This new version makes blind angle also
work consistently for all directions.