TreeItem: fix broken height on collapsing/uncollapsing nested treeitem lists
Idea is to not animate multiple lists at once for peformance and accuracy reasons. So on toggle the list element is set to height:auto to measure the resulting height. Depending on collapse or uncollapse the start height is set. in the next render cycle the final height is set so css animation starts. When display_state is set to show the height is removed from the style so the element reacts as 'auto' again. This way only the animating list item needs to be animated in order to define the height of all parent lists automatically.